Username and password label bug dropping fixed
