+++ 42-transcendence/client/src/components/ProtectPage.tsx
@@ -3,10 +3,6 @@ import { useRouter } from "next/router";
 import { ComponentType,useEffect } from "react";
 import  NextComponentType  from "next";
 
-/* The  HOC technique is being used to protect the pages  */ 
-/* <P extends object> : generic type.
-- Our component type is componentType<P>, P are the props passed */ 
-
 
 async function getUser(email:string) {
