--- 42-transcendence/client/src/components/profile/ProfileFriends.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react';
-
-function ProfileFriend() {
-    return ( <div className='profile-friend profile-container'>
-        <a href=""><img src="" alt="first friend"></img></a>
-        <a href=""><img src="" alt="second friend"></img></a>
-        <a href=""><img src="" alt="third friend"></img></a>
-    </div> );
-}
-
-export default ProfileFriend;
