--- 42-transcendence/client/src/components/profile/ProfileSettings.tsx
@@ -1,21 +1,0 @@
-import React from "react";
-
-
-
-
-
-function ProfileSettings() {
-    return ( <div className="profile settings">
-                <img src="" alt="profilePic"></img>
-                {/* <InputSettings for="ID" />
-                <InputSettings for="First Name" />
-                <InputSettings for="Last Name" />
-                <InputSettings for="Level" />
-                <InputSettings for="Rank" />
-                <InputSettings for="Phone Number" />
-                <InputSettings for="Date Of Birth" /> */}
-
-            </div>);
-}
-
-export default ProfileSettings;
