+++ 42-transcendence/client/src/components/profile/UpdateProfileComp.tsx
@@ -1,0 +1,9 @@
+
+
+function UpdateProfile() {
+    return (<div>
+        
+    </div>  );
+}
+
+export default UpdateProfile;
