+++ 42-transcendence/client/src/styles/globals.css
@@ -10,6 +10,7 @@ html {
   background: linear-gradient(90deg, rgba(44,59,124,1) 0%, rgba(13,23,67,1) 100%);
   font-family: 'Outfit'
 }
+
 .sh{
   text-shadow: 2px 2px 2px black;
 
@@ -23,5 +24,6 @@ html {
     margin-bottom:0;
   }
 }
+
 
 /******************************* profile css **********************************/
